Fix case where ~/.ssh/id_rsa.pub is a softlink
Some users may have the default id_rsa/id_rsa.pub be a softlink to one
of a set of rsa keys. In this case, the cp would result in a dangling
softlink in the container and the startup fails. Adding a -L to the cp
command results in copying whatever the softlink points to into the
container, which seems like the desired behaviour.

Signed-off-by: brian avery <brian.avery@intel.com>